Support `.tsx`
The extension `.tsx` is the typescript equivalent to `.jsx`. It should be supported as well! 

Right here I added it as an alias to `.jsx` but it would be better to give  `.tsx` an alternate color from `.jsx`

One thing to note:
`.js` icon is yellow.
`.ts` icon is **blue**.
`.jsx` icon is **blue**.
`.tsx` icon is ???

Perhaps `.jsx` become yellow, `.tsx` becomes blue? Seems more idiomatic to me.