fix(vsce): fix the vsce version
New 1.93.0 version makes the publication fail with an error saying that the The specified icon 'extension/images/logo.png' wasn't found in the extension.

I looked into vsce issues and there was one related which was closed without any answer in 2018. I'll get back to it at some point, but for the moment I will merge this to unblock our release.