Add .devcontainer.json to devcontainer
Apparently VS Code supports both [`devcontainer/devcontainer.json` and `.devcontainer.json` in the root of the project](https://code.visualstudio.com/docs/remote/create-dev-container#_create-a-devcontainerjson-file), so I added `.devcontainer.json` to the `extensions` field for the `devcontainer` icon.

Assuming that this change is okay, I'm unsure if I should update the Wiki or if that's something you do (or if it's automated). Just let me know!