fix(defaults): allow to use bundled icons for default files/folders
Adds a new property to the `IDefaultExtension` interface to allow the usage of bundled icons.

This basically allows anyone to use not only custom icons but use the bundled icons as default folders and files. Note that only folder icons can be assigned to default folder and root_folder.

fix Configuration for default (root-)folder icon #3103