feat(icon): support the go.mod and go.work languages
This adds support for the go.mod and go.work languages registered by
https://marketplace.visualstudio.com/items?itemName=golang.Go

The file names were already supported. The icon is now also visible in
the language selector.

There is also the related go.sum language, but the matching filename is
sometimes associated with the go.mod icon, and sometimes with the
go.work icon.