chore: update to Node.js 16.14.2 in wpil workflow
Before, Node.js 14 was used. A step would update npm to the latest
version. However, npm 10, the latest version of npm at the moment of
writing, doesn’t support Node.js 14.

Instead of using such an old version of Node.js, the Node.js version was
updated to 16.14.2. This is already EOL, so ideally an even more recent
version would be used. The npm shipped with this Node.js version is
probably recent enough, so the npm update step was removed.