fix: update to npm 9 in wpil workflow
The latest version of npm (10 at the moment of writing) is not compatible
with the Node.js version used. This caused the wpil workflow to always
fail. This error is marked by the `continue-in-error` option. The fix is
to pin the npm version to npm 9, latest npm version that’s still
compatible with Node.js 14.