feat(icon): add icon for fly.toml
The `fly.toml` file is a configuration file for https://fly.io. The icon
was taken from their brand page on https://fly.io/docs/about/brand/ and
adheres to their guidelines. Unfortunately it does somewhat exceeds the
file size limit.

Their own website uses a variation of the icon as a favicon with just
the balloon and a purple background. We might consider that instead.