feat(icon): add just icon
The `just` language ID is registered by
https://marketplace.visualstudio.com/items?itemName=skellock.just and
https://marketplace.visualstudio.com/items?itemName=kokakiwi.vscode-just.

I manually recreated the icon to match the favicon of
https://just.systems by overlaying it.

https://plugins.jetbrains.com/plugin/18658-just provides a verion of the
icon where the holes are transparent. To me the white dots on a black
circle look seem pretty iconic.

It also looks pretty good in dark mode in my opinion, despite the outer
circle being black.

Closes #2679