feat(zed): create a Zed extension manifest (#3865)
This PR extends the build proces to create a Zed extension manifest.

Zed icon extensions don't allow dynamic recreation of the manifest so there's no way to offer presets in the way we do in VS Code.

In order to mitigate that, the generated manifest will contain several theme variations:

- Normal icons
- Official icons
- Angular icons
- NestJs icons

Another interesting thing to consider is that Zed icon extensions don't support language ids. In order to make it work for Zed, some changes were introduced in the `languages.ts` file. The `defaultExtension` property has been removed as it was not used, and it has been replaced by a more complex property (optional) that includes extensions and file names.

The generated manifest will be used in the Zed extension repo: https://github.com/vscode-icons/vcode-icons-zed.

It also fixes an issue introduced in the last release affecting the `package.json`.